requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
qdrant-client==1.7.3
openai==1.12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
PyPDF2==3.0.1
pdfplumber==0.9.0
markdown==3.5.1
python-dotenv==1.0.0
httpx==0.25.2